feat: fully implement optimized Git target init and fix rule engine
- Implemented high-performance recursive Git config loading in `src/targets/gitConfig.ts`.
- Introduced `createGit()` for state isolation and persistent caching via `WeakMap`.
- Optimized core rule matching and path normalization (`unixify`) for "blazing" speed.
- Corrected `invert` mode and prioritized local overrides in `ruleTestSync`.
- Set `configVersion: 1` in `bun.lock` for isolated linker support.
- Achieved ~170µs Git init and verified with 122 passing tests.
