+++ src/targets/jsr.ts
@@ -34,6 +34,7 @@ const internal: Rule[] = [
  * @since 0.6.0
  */
 export const JSR: Target = <Target>{
+	configVersion: 1,
 	extractors,
 	ignores: ruleTest,
 	init({ fs, cwd }, cb) {
