+++ src/targets/testGit.test.ts
@@ -3,7 +3,7 @@ import type { NestedDirectoryJSON } from "memfs"
 import { test, describe } from "bun:test"
 
 import { testScan, type PathHandlerOptions } from "../testScan.test.js"
-import { Git as target } from "./git.js"
+import { Git, createGit } from "./git.js"
 
 async function testGit(
 	done: () => void,
@@ -11,6 +11,7 @@ async function testGit(
 	handler: ((o: PathHandlerOptions) => void | Promise<void>) | string[],
 ) {
 	try {
+		const target = createGit()
 		await testScan(done, tree, handler, { target })
 	} catch (error) {
 		throw new Error("Error while testing Git", { cause: error })
