+++ src/targets/vsce.ts
@@ -71,6 +71,7 @@ const internal: Rule[] = [
  * @since 0.6.0
  */
 export const VSCE: Target = <Target>{
+	configVersion: 1,
 	extractors,
 	ignores: ruleTest,
 	init({ fs, cwd }, cb) {
