+++ src/targets/yarn.ts
@@ -76,6 +76,7 @@ const internal: Rule[] = [
  * @since 0.6.0
  */
 export const Yarn: Target = <Target>{
+	configVersion: 1,
 	extractors,
 	ignores: ruleTest,
 	init({ fs, cwd }, cb) {
